添加imagenet1K数据集配置；修改segmae的masking token成为cls token；修改masking ratio成为batch内平均；

+++ mmpretrain/datasets/transforms/processing.py
@@ -213,7 +213,7 @@ class SegRandomResizedCrop(BaseTransform):
         seg_map = results['gt_seg_mask']
         seg_map = mmcv.imcrop(seg_map, bboxes=np.array([xmin, ymin, xmax, ymax]))
         results['gt_seg_mask'] = mmcv.imresize(
-            seg_map,
+            seg_map.astype(np.uint8),
             tuple(self.size[::-1]),
             interpolation='nearest',
             backend=self.backend)
